Interfaces

Documentation for Interfaces.

Interfaces.Interface

Interfaces.components

Interfaces.description

Interfaces.implements

Interfaces.optional_keys

Interfaces.test

Interfaces.test_objects

Interfaces.@implements

Interfaces.@interface

Interfaces.Interface — Type

Interface{Components}

Abstract supertype for all Interfaces.jl interfaces.

Components is an Tuple of Symbol.

Interfaces.components — Function

components(::Type{<:Interface})

Returns the components of the interface, as a NamedTuple of NamedTuple.

Interfaces.description — Function

description(::Type{<:Interface})

Returns a String description of an interface.

Interfaces.implements — Function

implements(::Type{<:Interface}, obj)
implements(::Type{<:Interface{Options}}, obj)

Returns whether an object implements an interface, as a Bool.

obj can be an be an object or a Type.

Options can be a Symbol or a Tuple of Symbol passed to the type parameter of the Interface, to check if optional interfaces are implemented by the obj.

Without specifying Options, the return value specifies that at least all the mandatory components of the interace are implemented.

Interfaces.optional_keys — Function

optional_keys(T::Type{<:Interface}, obj::Type)

Get the keys for the optional components of an Interface, as a tuple os Symbol.

Interfaces.test — Method

test(::Type{<:Interface}, obj)

Test if an interface is implemented correctly for an object, returning true or false.

If no interface type is passed, Interfaces.jl will find all the interfaces available and test them.

Interfaces.test_objects — Function

test_objects(::Type{<:Interface}, ::Type)

Return the test object for an Interface and type.

Interfaces.@implements — Macro

@implements(interface, objtype, obj)
@implements(dev, interface, objtype, obj)

Declare that an interface implements an interface, or multipleinterfaces.

Also pass an object or tuple of objects to test it with.

The macro can only be used once per module for any one type. To define multiple interfaces a type implements, combine them in square brackets.

Passing the keyword dev as the first argument lets us show test output during development. Do not use dev in production code, or output will appear during package precompilation.

Example

Here we implement the IterationInterface for Base julia, indicating with (:indexing, :reverse) that our object can be indexed and works with Iterators.reverse:

using BaseInterfaces
@implements BaseInterfaces.IterationInterface{(:indexing,:reverse)} MyObject MyObject([1, 2, 3])

Interfaces.@interface — Macro

@interface(interfacename, components, [description])

Define an interface.

@interface MyInterface (
    mandatory = (
        length = x -> length(x) = prod(size(x)),
        ndims = x -> ndims(x) = length(size(x)),
    ),
    optional = (;)
) "A description of the interface"