Interfaces

Documentation for Interfaces.

Interfaces.Interface

Interfaces.components

Interfaces.implements

Interfaces.test

Interfaces.@document

Interfaces.@implements

Interfaces.@interface

Interfaces.Interface — Type

Interface{Components}

Abstract supertype for all interfaces.

Components is an NTuple of Symbol.

Interfaces.components — Function

components(T::Interface)

Returns the components of interface tests, as a NamedTuple of Tuple of functions or functors.

Interfaces.implements — Method

(T::Interface)

Returns whether an object implements and interface, as a Bool.

Interfaces.test — Method

test(obj)
test(::Type{<:Interface}, obj)

Test if an interface is implemented correctly for an object, returning true or false.

If no interface type is passed, Interfaces.jl will find all the interfaces available and test them.

Interfaces.@document — Macro

@document

Macro to insert interface documentation into docs for objects That implement them.

Example

Interfaces.@document MyModule SomInterface

Interfaces.@implements — Macro

@implements(interface, objtype, obj)

Declare that an interface implements an interface, or multipleinterfaces.

Also pass an object or tuple of objects to test it with.

The macro can only be used once per module for any one type. To define multiple interfaces a type implements, combine them in square brackets.

Example

@implements BaseInterfaces.Iteration{(:indexed,)} MyArray MyArray([1, 2, 3])

Interfaces.@interface — Macro

@interface

Define an interface.

@interface MyInterface begin
    mandatory = (
        length = x -> length(x) = prod(size(x)),
        ndims = x -> ndims(x) = length(size(x)),
    ),
    optional = (;)
end